+++ catalog/app/containers/Bucket/Queries/Athena/Athena.tsx
@@ -19,6 +19,19 @@ import Results from './Results'
 import History from './History'
 import Workgroups from './Workgroups'
 
+interface QuerySelectSkeletonProps {
+  className?: string
+}
+
+function QuerySelectSkeleton({ className }: QuerySelectSkeletonProps) {
+  return (
+    <div className={className}>
+      <Skeleton height={24} width={128} animate />
+      <Skeleton height={48} mt={1} animate />
+    </div>
+  )
+}
+
 const useAthenaQueriesStyles = M.makeStyles((t) => ({
   form: {
     margin: t.spacing(3, 0, 0),
@@ -28,14 +41,12 @@ const useAthenaQueriesStyles = M.makeStyles((t) => ({
 interface QueryConstructorProps {
   bucket: string
   className?: string
-  queryExecutionId?: string
-  initialValue?: string
+  initialValue?: requests.athena.QueryExecution
   workgroup: requests.athena.Workgroup
 }
 
 function QueryConstructor({
   bucket,
-  queryExecutionId,
   className,
   initialValue,
   workgroup,
@@ -44,18 +55,25 @@ function QueryConstructor({
   const [prev, setPrev] = React.useState<requests.athena.QueriesResponse | null>(null)
   const data = requests.athena.useQueries(workgroup, prev)
   const classes = useAthenaQueriesStyles()
-  const [value, setValue] = React.useState<string | null>(initialValue || null)
-  const handleQueryBodyChange = React.useCallback((v: string) => {
-    setValue(v)
-    setQuery(null)
-  }, [])
+  const [value, setValue] = React.useState<requests.athena.QueryExecution | null>(
+    initialValue || null,
+  )
   const handleNamedQueryChange = React.useCallback(
     (q: requests.athena.AthenaQuery | null) => {
       setQuery(q)
-      setValue(q?.body || null)
+      setValue((x) => ({
+        ...x,
+        query: q?.body,
+      }))
     },
     [],
   )
+
+  const handleChange = React.useCallback((x: requests.athena.QueryExecution) => {
+    setValue(x)
+    setQuery(null)
+  }, [])
+
   return (
     <div className={className}>
       {data.case({
@@ -72,22 +90,26 @@ function QueryConstructor({
           </Section>
         ),
         Err: makeAsyncDataErrorHandler('Select query'),
-        _: () => (
-          <>
-            <Skeleton height={24} width={128} animate />
-            <Skeleton height={48} mt={1} animate />
-          </>
-        ),
+        _: () => <QuerySelectSkeleton />,
       })}
       <QueryEditor.Form
         bucket={bucket}
         className={classes.form}
-        queryExecutionId={queryExecutionId}
-        workgroup={workgroup}
-        onChange={handleQueryBodyChange}
+        onChange={handleChange}
         value={value}
+        workgroup={workgroup}
       />
     </div>
+  )
+}
+
+function QueryConstructorSkeleton() {
+  const classes = useStyles()
+  return (
+    <>
+      <QuerySelectSkeleton className={classes.section} />
+      <QueryEditor.Skeleton className={classes.section} />
+    </>
   )
 }
 
@@ -322,15 +344,22 @@ interface AthenaMainProps {
 
 function AthenaMain({ bucket, workgroup }: AthenaMainProps) {
   const classes = useStyles()
+  const data = requests.athena.useDefaultQueryExecution()
   return (
     <div className={classes.content}>
-      <QueryConstructor
-        bucket={bucket}
-        className={classes.section}
-        key={workgroup}
-        workgroup={workgroup}
-      />
-
+      {data.case({
+        Ok: (queryExecution) => (
+          <QueryConstructor
+            bucket={bucket}
+            className={classes.section}
+            key={workgroup}
+            workgroup={workgroup}
+            initialValue={queryExecution}
+          />
+        ),
+        Err: makeAsyncDataErrorHandler('Default catalog and database'),
+        _: () => <QueryConstructorSkeleton />,
+      })}
       <HistoryContainer
         bucket={bucket}
         className={classes.section}
@@ -356,20 +385,11 @@ function AthenaExecution({ bucket, workgroup, queryExecutionId }: AthenaExecutio
           <QueryConstructor
             bucket={bucket}
             className={classes.section}
-            queryExecutionId={queryExecutionId}
-            initialValue={value?.queryExecution?.query}
+            initialValue={value?.queryExecution}
             workgroup={workgroup}
           />
         ),
-        _: () => (
-          <>
-            <div className={classes.section}>
-              <Skeleton height={24} width={128} animate />
-              <Skeleton height={48} mt={1} animate mb={3} />
-            </div>
-            <QueryEditor.Skeleton className={classes.section} />
-          </>
-        ),
+        _: () => <QueryConstructorSkeleton />,
       })}
 
       {results.data.case({